Remove deprecated packages and data sources
<dataframe>.append(<dataframe>) is no longer supported in pandas.
Use pd.concat([df1, df2]) instead.

Minor corrections to how to call pip from a notebook (use % instead of !).

Minor correction to package isntallation (sci-kit learn).

Remove analysis of 2016 census data. The source/link is dead. The data was
already old when this project was created, so it's not much of a loss.